vk_auction_bot/requirements.txt
vkbottle==4.3.14
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
aiohttp==3.9.5
